split off the evaluation help into separate program
